pyproject.toml
[project]
name = "nonebot-plugin-deepseek"
version = "0.1.8"
description = "NoneBot DeepSeek 插件。接入 DeepSeek 模型，提供智能对话与问答功能"
authors = [
    { name = "KomoriDev ", email = "mute231010@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
keywords = []

requires-python = ">=3.9"
dependencies = [
    "nonebot2>=2.4.1",
    "nonebot-plugin-alconna>=0.55.0",
    "httpx[http2]>=0.28.1",
    "beautifulsoup4>=4.12.3",
    "nonebot-plugin-localstore>=0.7.3",
    "cli-lite>=0.10.2",
]

[project.entry-points.nb_scripts]
deepseek = "nonebot_plugin_deepseek.cli.__main__:main"


[project.urls]
homepage = "https://github.com/KomoriDev/nonebot-plugin-deepseek"
repository = "https://github.com/KomoriDev/nonebot-plugin-deepseek"

[project.optional-dependencies]
image = [
    "nonebot-plugin-htmlrender>=0.5.1",
]
adapters = [
    "nonebot-adapter-onebot>=2.4.6",
]
fastapi = [
    "nonebot2[fastapi]>=2.4.1",
]
test = [
    "pytest>=8.3.4",
    "nonebug>=0.4.3",
    "pytest-asyncio>=0.25.3",
]

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.6.1",
    "isort>=5.13.2",
    "pre-commit>=3.7.1",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_deepseek"]


[tool.isort]
profile = "black"
line_length = 120
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.ruff.lint.isort]
length-sort = true
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "standard"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
